Renamed the packages and modules now that we have the testjavafx.org domain name.
