+++ internal-test/src/test/java/NodeQueryTest.java
@@ -11,8 +11,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Licence for the specific language governing permissions and limitations under the Licence.
  */
-import com.eponymouse.testjavafx.junit4.ApplicationTest;
-import javafx.animation.Animation;
+import org.testjavafx.junit4.ApplicationTest;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.css.Styleable;
