+++ internal-test/src/test/java/TypingTest.java
@@ -11,8 +11,8 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Licence for the specific language governing permissions and limitations under the Licence.
  */
-import com.eponymouse.testjavafx.FxThreadUtils;
-import com.eponymouse.testjavafx.junit4.ApplicationTest;
+import org.testjavafx.FxThreadUtils;
+import org.testjavafx.junit4.ApplicationTest;
 import javafx.scene.Scene;
 import javafx.scene.control.TextField;
 import javafx.scene.input.KeyCode;
