+++ testjavafx-core/src/main/java/org/testjavafx/FxRobotInterface.java
@@ -11,9 +11,9 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Licence for the specific language governing permissions and limitations under the Licence.
  */
-package com.eponymouse.testjavafx;
+package org.testjavafx;
 
-import com.eponymouse.testjavafx.node.NodeQuery;
+import org.testjavafx.node.NodeQuery;
 import javafx.geometry.Point2D;
 import javafx.scene.Node;
 import javafx.stage.Window;
