+++ testjavafx-core/src/main/java/org/testjavafx/FxRobotInterfaceWindow.java
@@ -11,7 +11,7 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the Licence for the specific language governing permissions and limitations under the Licence.
  */
-package com.eponymouse.testjavafx;
+package org.testjavafx;
 
 import com.google.common.collect.ImmutableList;
 import javafx.stage.PopupWindow;
